edit settings.py

+++ check_phone/models.py
@@ -22,7 +22,7 @@ class Requests(models.Model):
 class Request(models.Model):
     id = models.BigIntegerField(primary_key=True)
     requests_id = models.IntegerField()
-    phone = models.IntegerField()
+    phone = models.BigIntegerField()
     hlr_status = models.CharField(max_length=100)
     hlr_status_code = models.IntegerField(null=True, blank=True)
